Handle circular subdelegation edge cases (#121)
* get a list of direct and indirect delegators

* refactor FE to fetch all delegators

* remove duplicate actions

* remove console log
