+++ src/components/Delegates/DelegateCard/AdvancedDelegateButton.tsx
@@ -6,7 +6,6 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { Delegation } from "@/app/api/common/delegations/delegation";
 import { useOpenDialog } from "@/components/Dialogs/DialogProvider/DialogProvider";
 import {
   fetchVotingPowerForSubdelegation,
@@ -21,7 +20,7 @@ export function AdvancedDelegateButton({
   delegators,
 }: {
   delegate: DelegateChunk;
-  delegators: Delegation[] | null;
+  delegators: string[] | null;
 }) {
   const { address } = useAccount();
 
@@ -31,10 +30,7 @@ export function AdvancedDelegateButton({
   const isDisabled =
     !!address &&
     !!(
-      isSelfDelegation ||
-      delegators
-        ?.map((delegator) => delegator.from.toLowerCase())
-        .includes(delegate.address?.toLowerCase())
+      isSelfDelegation || delegators?.includes(delegate.address?.toLowerCase())
     );
 
   return (
@@ -51,7 +47,10 @@ export function AdvancedDelegateButton({
         <TooltipContent>
           <p>
             Delegating back to{" "}
-            {isSelfDelegation ? "yourself" : "your delegator"} is not supported
+            {isSelfDelegation
+              ? "yourself"
+              : "your direct or indirect delegator"}{" "}
+            is not supported
           </p>
         </TooltipContent>
       </Tooltip>
