+++ src/components/Votes/VoterHoverCard.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { useEffect, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { VStack, HStack } from "@/components/Layout/Stack";
 import { DelegateProfileImage } from "../Delegates/DelegateCard/DelegateProfileImage";
 import { useRouter } from "next/navigation";
@@ -12,63 +12,46 @@ import { AdvancedDelegateButton } from "../Delegates/DelegateCard/AdvancedDelega
 import { DelegateButton } from "../Delegates/DelegateCard/DelegateButton";
 import { Delegate } from "@/app/api/common/delegates/delegate";
 import { DelegateStatement } from "@/app/api/common/delegateStatement/delegateStatement";
-import { Delegation } from "@/app/api/common/delegations/delegation";
+import { fetchDelegate, fetchDelegateStatement } from "@/app/delegates/actions";
 
 interface Props {
   address: string;
-  fetchDelegate: (addressOrENSName: string) => Promise<any>;
-  fetchDelegateStatement: (addressOrENSName: string) => Promise<any>;
-  fetchBalanceForDirectDelegation: (
-    addressOrENSName: string
-  ) => Promise<string>;
-  fetchVotingPowerForSubdelegation: (
-    addressOrENSName: string
-  ) => Promise<string>;
-  checkIfDelegatingToProxy: (addressOrENSName: string) => Promise<boolean>;
-  fetchCurrentDelegatees: (addressOrENSName: string) => Promise<any>;
-  fetchDirectDelegatee: (addressOrENSName: string) => Promise<any>;
-  getProxyAddress: (addressOrENSName: string) => Promise<string>;
   isAdvancedUser: boolean;
-  delegators: Delegation[] | null;
+  delegators: string[] | null;
 }
 
 export default function VoterHoverCard({
   address,
-  fetchDelegate,
-  fetchDelegateStatement,
-  fetchBalanceForDirectDelegation,
-  fetchVotingPowerForSubdelegation,
-  checkIfDelegatingToProxy,
-  fetchCurrentDelegatees,
-  fetchDirectDelegatee,
-  getProxyAddress,
   isAdvancedUser,
   delegators,
 }: Props) {
   const router = useRouter();
   // delegateStatement is loaded separately to avoid delays in loading the hover card
   const [delegateStatement, setDelegateStatement] =
-    useState<DelegateStatement>();
+    useState<DelegateStatement | null>();
   // full delegate object is required for the delegate button, that can appear later
   const [delegate, setDelegate] = useState<Delegate>();
 
   const { isConnected } = useAgoraContext();
   const { address: connectedAddress } = useAccount();
 
-  const fetchDelegateAndSet = async (addressOrENSName: string) => {
+  const fetchDelegateAndSet = useCallback(async (addressOrENSName: string) => {
     const delegate = await fetchDelegate(addressOrENSName);
     setDelegate(delegate);
-  };
+  }, []);
 
-  const fetchDelegateStatementAndSet = async (addressOrENSName: string) => {
-    const delegateStatement = await fetchDelegateStatement(addressOrENSName);
-    setDelegateStatement(delegateStatement);
-  };
+  const fetchDelegateStatementAndSet = useCallback(
+    async (addressOrENSName: string) => {
+      const delegateStatement = await fetchDelegateStatement(addressOrENSName);
+      setDelegateStatement(delegateStatement);
+    },
+    []
+  );
 
   useEffect(() => {
     fetchDelegateStatementAndSet(address);
     fetchDelegateAndSet(address);
-  }, []);
+  }, [fetchDelegateAndSet, fetchDelegateStatementAndSet, address]);
 
   let truncatedStatement = "";
   if (delegate?.statement?.payload) {
